rename file and move to the button folder
