+++ packages/components/src/shared/src/index.ts
@@ -32,7 +32,6 @@ export * from "./useId";
 export * from "./useKeyboardNavigation";
 export * from "./useHasChild";
 export * from "./useIsInitialRender";
-export * from "./useDisableInteraction";
 export * from "./useDisposables";
 export * from "./useForceRender";
 export * from "./useResizeObserver";
